Worst rat infestation solved.

+++ include/permonst.h
@@ -65,7 +65,7 @@ struct permonst {
 	unsigned long mconveys;             /* conveyed by eating */
     unsigned long mflags1,      /* boolean bitflags */
         mflags2, mflags3;       /* more boolean bitflags */
-    uchar difficulty;		/* toughness (formerly from  makedefs -m) */
+    int difficulty;		/* toughness (formerly from  makedefs -m) */
 #ifdef TEXTCOLOR
     uchar mcolor; /* color to use */
 #endif
